refactor(xmlrpc): extract fault tests to sibling file
